requirements.txt
hydra-core==1.3.2
pandas==2.1.4
pytest==7.4.4
scikit_learn==1.3.2
timm==0.9.12
pytest
torch==2.1.2
torchvision==0.16.2
wandb==0.16.2
matplotlib==3.8.2